src/test/java/com/joggingthoughts/controller/UserControllerTests.java
package com.joggingthoughts.controller;

import com.joggingthoughts.domain.UserEntity;
import com.joggingthoughts.service.UserService;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.Mockito;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;


import java.util.Arrays;
import java.util.List;

import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
import static org.hamcrest.Matchers.*;

@WebMvcTest(controllers = UserController.class)
public class UserControllerTests {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private UserService userService;

    @BeforeEach
    public void setUp(){
        List<UserEntity> users = Arrays.asList(
                new UserEntity("Jane Doe", "example@contesto.com"),
                new UserEntity("John Doe", "johndoe@example.com")
        );

        Mockito.when(userService.getAllUsers()).thenReturn(users);
    }

    @Test
    public void testGetAllUsers() throws Exception{

        mockMvc.perform(get("/api/users")
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$", hasSize(2)))
                .andExpect(jsonPath("$[0].name", is("Jane Doe")))
                .andExpect(jsonPath("$[1].name", is("John Doe")));

    }
}
